♻️ re arrange useful testing code for provider tests
