+++ tests/Providers/Qr/IQRCodeProviderTest.php
@@ -5,26 +5,11 @@ namespace Tests\Providers\Qr;
 use PHPUnit\Framework\TestCase;
 use RobThree\Auth\TwoFactorAuth;
 use RobThree\Auth\TwoFactorAuthException;
+use RobThree\Auth\Providers\Qr\HandlesDataUri;
 
 class IQRCodeProviderTest extends TestCase
 {
-    /**
-     * @param string $datauri
-     *
-     * @return null|array
-     */
-    private function DecodeDataUri($datauri)
-    {
-        if (preg_match('/data:(?P<mimetype>[\w\.\-\/]+);(?P<encoding>\w+),(?P<data>.*)/', $datauri, $m) === 1) {
-            return array(
-                'mimetype' => $m['mimetype'],
-                'encoding' => $m['encoding'],
-                'data' => base64_decode($m['data'])
-            );
-        }
-
-        return null;
-    }
+    use HandlesDataUri;
 
     /**
      * @return void
